Extract catalog resolution surfaces
